pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "flask-sqlalchemy>=3.1.1,<4.0.0",
    "pydantic (>=2.12.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",

]
[tool.poetry]
package-mode = false